fixing memory leak
The controller does not have enough ram to reliably increase the subcode count.
As a result the probing commands have been moved to M460-M465. This also better matches the Linuxcnc standard.

The second memory leak was a result of failing to delete after calling Gcode new, leading to unfreed references