set tool length by offset M493.3 H{#}
M493.4 H5 - move the tip of the tool so it is 5mm above the current WCS Z 0, then run this command.

I accomplish this by lowering the bit until a 5mm endmill on the Z0 surface cannot roll underneath the bit. Then I raise the z axis by small increments until the endmill can just pass underneath the bit with a tiny amount of resistance.

This command will then autocalculate the TLO for current tool. When complete the tool will be located at 5mm in Z.