Improved build scripts and cleanup
New build scripts documented in DEVELOPER.md dramatically simplify
working with different GCC toolchains.

- build/gcc.sh manages GCC toolchain download and command execution
- build/build.sh runs multi-core builds using build/gcc.sh
- mri/gdb.sh updated to use build/gcc.sh to find GDB
- moved dev docs to DEVELOPER.md
- added static analysis documentation to DEVELOPER.md
- added TODO.md to track interesting tasks for the future

Smoothieware build files have not been removed as yet out of an
abundance of caution, just moved to attic so that users don't attempt
to use them. This results in a scary looking commit, but it's just file
moves. The reality is that these files are either completely irrelevant
to the Carvera or so significantly out of date as to be the same, so
they should be removed in the near future.